refactor: 자판기 관리 메뉴 생성 및 클릭 이벤트 view로 이동
